Added auth middleware in crud operation and also added comment of each controller

+++ backend/controllers/course.controller.js
@@ -2,13 +2,27 @@ const asyncHandler = require("express-async-handler")
 const Course = require("../models/course.model")
 const User = require("../models/user.model")
 
-
-const getUserCourse = asyncHandler( async(request, response) => {
+//@desc Get all courses of all users
+//@route GET /api/courses
+//@access public
+const getAllCourses = asyncHandler( async(request, response) => {
     const courses = await Course.find()
 
     response.status(200).json({ courses })
 })
 
+//@desc Get all course of authorized user
+//@route GET /api/courses
+//@access private
+const getUserCourses = asyncHandler( async(request, response) => {
+    const courses = await Course.find({ user: request.user._id})
+
+    response.status(200).json({ courses })
+})
+
+//@desc create course of authorized user
+//@route POST /api/courses
+//@access private
 const createCourse = asyncHandler( async(request, response) => {
     const { courseName } = request.body
 
@@ -18,12 +32,16 @@ const createCourse = asyncHandler( async(request, response) => {
     }
 
     const newCourse = await Course.create({
-        courseName
+        courseName,
+        user: request.user._id
     })
 
     response.status(201).json({ newCourse })
 })
 
+//@desc update course of authorized user
+//@route PUT /api/courses/:id
+//@access private
 const updateCourse = asyncHandler( async(request, response) => {
     const { id } = request.params
 
@@ -33,10 +51,21 @@ const updateCourse = asyncHandler( async(request, response) => {
         throw new Error("Course not found")
     }
 
+    const user = await User.findById(request.user._id)
+    // if(course.user.toString() !== user._id){
+    if(!course.user.equals(user._id)){
+        response.status(401)
+        throw new Error("Not Authorized")
+    }
+
     const updatedCourse = await Course.findByIdAndUpdate(id, request.body, { new: true})
     response.status(200).json(updatedCourse)
 })
 
+
+//@desc delete course of authorized user
+//@route DELETE /api/courses/:id
+//@access private
 const deleteCourse = asyncHandler( async(request, response) => {
     const { id } = request.params
     
@@ -46,12 +75,20 @@ const deleteCourse = asyncHandler( async(request, response) => {
         throw new Error("Course not found")
     }
 
+    const user = await User.findById(request.user._id)
+    // if(course.user.toString() !== user._id){
+    if(!course.user.equals(user._id)){
+        response.status(401)
+        throw new Error("Not Authorized")
+    }
+
     await Course.findByIdAndDelete(id)
     response.status(200).json({message: `course of ${id} deleted`})
 })
 
 module.exports = {
-    getUserCourse,
+    getAllCourses,
+    getUserCourses,
     createCourse,
     updateCourse,
     deleteCourse
